Some Basic codes
nc_inspect for inspecting the NetCDF4 file. To see the variables, units etc in the file.
nc_visual for visualizing the NetCDF4 file as a global map. 
ensembles_compare.py : The code used to plot the ensemble members' performance. To check if the forecasts diverging as longer the forecast horizon.